% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_models.R
\name{convergence_check}
\alias{convergence_check}
\title{Convergence checks}
\usage{
convergence_check(model_output)
}
\description{
Returns the model run time, the number of parameters with Rubin-Gelman
statistics greater than 1.1 (indicating that they have not converged),
and the number of effective samples of the worst sampled parameter.
}
