% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figures.R
\name{correlation_heatmap_figure}
\alias{correlation_heatmap_figure}
\title{Correlation heatmap figure}
\usage{
correlation_heatmap_figure(model = "m3")
}
\description{
Creates an S x S heatmap of covariances between species. Red is negative,
blue is positive.
}
