% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figures.R
\name{covariance_density_figure}
\alias{covariance_density_figure}
\title{Covariance change between treatments}
\usage{
covaraiance_density_figure("m6",
species = c("Avena.fatua", "Bromus.diandrus", "Acetosella.vulgaris))
}
\description{
Creates density plots of covariances of selected species to compare change
in interaction strength between treatments.
}
