% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figures.R
\name{covariance_heatmap_figure}
\alias{covariance_heatmap_figure}
\title{Covariance heatmap figure}
\usage{
covariance_heatmap_figure(model = "m3")
}
\description{
Creates an S x S heatmap of covariances between species. Red is negative,
blue is positive.
}
