% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figures.R
\name{create_figures}
\alias{create_figures}
\title{Create figures}
\usage{
create_figures_models(model = "m0", fig = 1)
}
\arguments{
\item{path}{Directory to save model outputs, defaults to /models.}

\item{figs}{numeric: 1-4, defaults to all.}

\item{model}{m0-m6, defaults to all.}
}
\description{
Creates a series of figures to aide the interpretation of model outputs.
The figures are model dependent and should only be used if the models have
converged correctly.
}
\details{
Figures include:
\enumerate{
\item{a highest-density interval summary of logistic regression (m0)}
\item{linear tobit predictions of species cover along a fertility gradient
(m1-3)}
\item{heatmaps of between species covariances (m2 & m3)}
\item{network diagrams of negative interactions between species (m1-3)}
\item{density plot of changes in negative interactions between treatments (m2 & m3)}
\item{trait-covariance correlation plots, varying by treatment (m3)}
}
}
