% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figures.R
\name{extract_pars}
\alias{extract_pars}
\title{Extract parameter estimates}
\usage{
pars <- extract_par(model_output, c("B"),
c("species", "covariate", "treatment"))
}
\arguments{
\item{fit}{output of a Stan model}

\item{pars}{vector of parameter names}

\item{index}{vector of labels for parameter indexes}
}
\description{
Stan output provides summaries of parameter estimates, this can be
faster than working directly with samples
}
