% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figures.R
\name{hdi_interval_figure}
\alias{hdi_interval_figure}
\title{Highest density interval figure}
\usage{
hdi_interval_figure(model = "m0")
}
\description{
Summarises posterior intervals in a whisker plot. Only defined for logistic
regression of the proportion of exotic species (m0).
}
