% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figures.R
\name{heatmap_network_figure}
\alias{heatmap_network_figure}
\title{Combine covariance heatmap and network}
\usage{
heatmap_network_figure(model, model_output)
}
\description{
Combine covariance heatmap and network
}
