% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figures.R
\name{impact_heatmap_figure}
\alias{impact_heatmap_figure}
\title{Impact heatmap figure}
\usage{
impact_heatmap_figure(model = "m3")
}
\description{
Creates an S x S heatmap of covariances between species. Red is negative,
blue is positive.
}
