% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figures.R
\name{interaction_network_figure}
\alias{interaction_network_figure}
\title{Interaction network figure}
\usage{
interaction_network_figure(model = "m6")
}
\description{
Filters between species covariance matrices for significant (non-zero)
negative interactions and displays them as a network. Links between species
are weighted by the magnitude of the interaction, but are only relative
within a given network
}
