% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{label}
\alias{label}
\title{Add corner label to plots}
\usage{
label(ggplot, label, x = 0.05)
}
\arguments{
\item{ggplot}{ggplot object}

\item{label}{string to label in top left corner}
}
\description{
Add corner label to plots
}
