% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figures.R
\name{linear_tobit_figure}
\alias{linear_tobit_figure}
\title{Linear tobit predictions figure}
\usage{
linear_tobit_figure(model = "m1")
}
\description{
Predicts the latent values of species abundance along a fertility gradient.
}
