% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_models.R
\name{load_model}
\alias{load_model}
\title{Load model}
\usage{
model_output <- load_models(model = "m1", path = "models/")
}
\arguments{
\item{model}{m0-m6, defaults to all.}

\item{path}{Directory to save model outputs, defaults to /models.}
}
\description{
Loads model output from .Rdata file.
}
