% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models.R
\name{logistic_regression}
\alias{logistic_regression}
\title{Logistic regression}
\usage{
logistic_regression(model = "m0", path = "models/")
}
\arguments{
\item{model}{"m0" only}

\item{path}{path to jags model (defaults to "models/")}
}
\description{
Run logit regression in JAGS
}
