% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models.R
\name{nested}
\alias{nested}
\title{Index nested factors}
\usage{
nested(plot, site)
}
\description{
Used for double-indexing in Stan models. Assigns numerical id of top-level
factors to each entry of lower-level factors.
}
