% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figures.R
\name{nonnative_dominance_figures}
\alias{nonnative_dominance_figures}
\title{Nonnative dominance}
\usage{
dominance_figures(model = "m0")
}
\description{
Plots mean regression lines for each treatment and annual change in
nonnative-fertility relationship following fencing.
}
