% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_models.R
\name{posterior_check}
\alias{posterior_check}
\title{Posterior check}
\usage{
posterior_check(model_output)
}
\description{
Runs posterior checks on model output by comparing the observed data to
posterior predictions under the same conditions. Presently, these are only
defined for tobit models (m1-m6) and include the root square mean error
(RMSE) of predictions after censoring, the total Euclidean distance between
all censored predictions and the data, the accuracy of predicted presences
and absences given as proportion of correct predictions, and the R-squared
statistic of predicted positive abundances against data.
}
