% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figures.R
\name{pred_cover}
\alias{pred_cover}
\title{Gather and predict cover}
\usage{
pred_cover(estimates, range = 2, coef = c("int", "fert"))
}
\arguments{
\item{estimates}{estimates of species coefficients}

\item{range}{range to predict over, default is (-2, 2)}

\item{coef}{two of c("int", "fert", "rain")}
}
\description{
Long dataframe of coefficents from extract_pars(), return predictions.
}
