% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models.R
\name{run_stan_model}
\alias{run_stan_model}
\title{Run Stan model}
\usage{
run_stan_model(model, path, ...)
}
\arguments{
\item{model}{one of c("m0", "m1", "m2", "m3")}

\item{...}{tuning parameters for Stan sampler}
}
\description{
Contains pre-processing and wraps \link{rstan} to run analyses using
adaptive Hamiltonian Monte Carlo. Sensible default settings are provided,
but can be tweaked as necessary.
}
