% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models.R
\name{subset_species_abundance}
\alias{subset_species_abundance}
\title{Generate proportional abundance lists}
\usage{
subset_species(cover, 0.9)
}
\description{
Selects the most abundant species in the dataset up to some poportional
threshold. Species are ordered alphabetically and given numeric ids.
}
\details{
90% of observed cover is a typical default.
}
