% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models.R
\name{subset_species_plots}
\alias{subset_species_plots}
\title{Generate proportional presence lists}
\usage{
subset_species(cover, 0.15)
}
\description{
Selects the most abundant species in the dataset by the proportion plots
a species was observed in. Discards species absent in one or more years.
}
\details{
15% of plots is a typical default.
}
