% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figures.R
\name{trait_correlation_figure}
\alias{trait_correlation_figure}
\title{Changes in covariance with relative trait values}
\usage{
trait_correlation_figure("m4", trait = "SLA")
}
\description{
Plots correlations between the pairwise differences in species traits and
between species covariances.
}
\details{
The trait dataset includes:
\enumerate{
\item{Leaf.length}
\item{Leaf.width}
\item{Max.height}
\item{SLA}
\item{Vegetative.height}
\item{Vegetative.width}
}
}
