% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figures.R
\name{extract_pars}
\alias{extract_pars}
\title{Extract parameter estimates}
\usage{
extract_pars(summary, pars, index)
}
\arguments{
\item{summary}{output of a Stan model summary}

\item{pars}{vector of parameter names}

\item{index}{vector of labels for parameter indexes}
}
\description{
Stan output provides summaries of parameter estimates, this can be faster than working directly with samples
}
\examples{
pars <- extract_par(fit = model_output$model_summary,
                          pars = c("B"),
                          index = c("species", "covariate", "treatment"))
}
